third_party/WebKit/LayoutTests/fast/events/popup-blocked-from-fake-user-gesture-expected.txt

Click Here
PASS win is non-null.
PASS win is undefined.
PASS successfullyParsed is true

TEST COMPLETE
